backend/requirements.txt
# Core
fastapi==0.104.0
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic>=2.0.0

# File handling
python-multipart==0.0.6
Pillow>=10.0.0,<12.0.0

# Gemini AI
google-genai>=0.2.2

# Firebase Auth & Firestore
firebase-admin>=6.0.0
google-cloud-firestore>=2.14.0
google-auth>=2.0.0
requests>=2.28.0

# Environment
python-dotenv>=1.0.0